requirements.txt

numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
ta>=0.7.0  # Technical Analysis library
gym>=0.21.0
stable-baselines3>=1.4.0
yfinance>=0.1.70
pandas-ta>=0.3.14b0
alpha_vantage>=2.3.1
websocket-client>=1.2.3
ccxt>=1.60.0  # For market data access
backtrader>=1.9.76.123
tensorflow>=2.8.0
keras>=2.8.0
tqdm>=4.62.0
joblib>=1.1.0
optuna>=2.10.0